# packages/schema/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Sentinel AI API
  description: |
    HTTP contract for Sentinel AI ingest and control plane.

    Trace payloads MUST conform to `packages/schema/jsonschema/v1`.

    **Implementation status:** Ingest, projects, runs, health, quarantine,
    derived metrics, evaluation, benchmarking, and regression/CI gate routes
    are implemented by `apps/server`.
  version: 0.1.0
  contact:
    name: Sentinel AI Maintainers
    license:
      name: Apache-2.0
      url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local self-hosted (future Compose default)
...
paths:
...
  /v1/projects/{project_id}/benchmark-suites:
    get:
      tags: [benchmarks]
      summary: List benchmark suites
      operationId: listBenchmarkSuites
      security:
        - ApiKeyAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Benchmark suite list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/BenchmarkSuite"
    post:
      tags: [benchmarks]
      summary: Register a versioned benchmark suite
      description: |
        Defines tasks, optional linked eval suite, and sweep dimensions.
        Sentinel does not execute agents; callers register completed runs as cells.
      operationId: createBenchmarkSuite
      security:
        - ApiKeyAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BenchmarkSuiteCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BenchmarkSuite"
  /v1/projects/{project_id}/benchmark-suites/{benchmark_id}/cells:
    get:
      tags: [benchmarks]
      summary: List benchmark cells
      operationId: listBenchmarkCells
      security:
        - ApiKeyAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: benchmark_id
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Cells
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/BenchmarkCell"
    post:
      tags: [benchmarks]
      summary: Register a matrix cell (link a completed run)
      operationId: registerBenchmarkCell
      security:
        - ApiKeyAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: benchmark_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [task_id, run_id]
              additionalProperties: false
              properties:
                task_id:
                  type: string
                run_id:
                  type: string
                dimensions:
                  type: object
                  additionalProperties: true
                environment:
                  type: object
                  additionalProperties: true
                cell_id:
                  type: string
                eval_job_id:
                  type: string
                run_eval:
                  type: boolean
                benchmark_version:
                  type: string
      responses:
        "201":
          description: Cell registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BenchmarkCell"
  /v1/projects/{project_id}/benchmark-suites/{benchmark_id}/leaderboard:
    get:
      tags: [benchmarks]
      summary: Matrix leaderboard for a benchmark suite
      operationId: getBenchmarkLeaderboard
      security:
        - ApiKeyAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: benchmark_id
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: query
          schema:
            type: string
        - name: baseline_agent_version
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Leaderboard report
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /v1/projects/{project_id}/benchmarks:
    post:
      tags: [benchmarks]
      summary: Build a benchmark comparison report
      description: |
        Aggregates registered cells into a leaderboard and optional pairwise
        deltas vs baseline_agent_version. Executes synchronously.
      operationId: createBenchmarkJob
      security:
        - ApiKeyAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BenchmarkJobCreate"
      responses:
        "202":
          description: Benchmark job finished (succeeded or failed status)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BenchmarkJob"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/projects/{project_id}/benchmarks/{job_id}:
    get:
      tags: [benchmarks]
      summary: Get benchmark comparison job
      operationId: getBenchmarkJob
      security:
        - ApiKeyAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Benchmark job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BenchmarkJob"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Sentinel-Api-Key
      description: |
        Optional in local single-user mode. Required for non-local profiles in later phases.
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              description: Machine-readable code (e.g. SCHEMA_VALIDATION_FAILED)
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    BenchmarkJobCreate:
      type: object
      additionalProperties: false
      properties:
        suite_id:
          type: string
          description: Alias for benchmark_id (OpenAPI compatibility)
        benchmark_id:
          type: string
        benchmark_version:
          type: string
        baseline_agent_version:
          type: string
    BenchmarkJob:
      type: object
      required: [id, status]
      additionalProperties: false
      properties:
        id:
          type: string
        project_id:
          type: string
        benchmark_id:
          type: string
        benchmark_version:
          type: string
        status:
          type: string
          enum: [queued, running, succeeded, failed]
        baseline_agent_version:
          type: string
          nullable: true
        report:
          type: object
          additionalProperties: true
        error_message:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
          nullable: true
    BenchmarkSuiteCreate:
      type: object
      required: [benchmark_id, version, name, definition]
      additionalProperties: false
      properties:
        benchmark_id:
          type: string
        version:
          type: string
        name:
          type: string
        description:
          type: string
        definition:
          type: object
          additionalProperties: true
    BenchmarkSuite:
      type: object
      required: [benchmark_id, version, name, definition]
      additionalProperties: false
      properties:
        benchmark_id:
          type: string
        version:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        definition:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
    BenchmarkCell:
      type: object
      required: [cell_id, task_id, run_id, dimensions, environment_fingerprint]
      additionalProperties: false
      properties:
        cell_id:
          type: string
        benchmark_id:
          type: string
        benchmark_version:
          type: string
        task_id:
          type: string
        run_id:
          type: string
        eval_job_id:
          type: string
          nullable: true
        dimensions:
          type: object
          additionalProperties: true
        environment_fingerprint:
          type: object
          additionalProperties: true
        metrics_snapshot:
          type: object
          additionalProperties: true
          nullable: true
        scores_summary:
          type: object
          additionalProperties: true
          nullable: true
        created_at:
          type: string
          format: date-time
...